Restructure hook files
