+++ assets/js/hooks/cell_editor/live_editor.js
@@ -3,8 +3,8 @@ import EditorClient from "./live_editor/editor_client";
 import MonacoEditorAdapter from "./live_editor/monaco_editor_adapter";
 import HookServerAdapter from "./live_editor/hook_server_adapter";
 import RemoteUser from "./live_editor/remote_user";
-import { replacedSuffixLength } from "../lib/text_utils";
-import { settingsStore } from "../lib/settings";
+import { replacedSuffixLength } from "../../lib/text_utils";
+import { settingsStore } from "../../lib/settings";
 
 /**
  * Mounts cell source editor with real-time collaboration mechanism.
