+++ assets/js/hooks/cell_editor/live_editor/elixir/on_type_formatting_edit_provider.js
