+++ assets/js/hooks/cell_editor/live_editor/hook_server_adapter.js
@@ -1,4 +1,4 @@
-import Delta from "../../lib/delta";
+import Delta from "../../../lib/delta";
 
 /**
  * Encapsulates logic related to sending/receiving messages from the server.
