+++ assets/js/hooks/confirm_modal.js
