+++ assets/js/hooks/editor_settings.js
