+++ assets/js/hooks/focus_on_update.js
