+++ assets/js/hooks/keyboard_control.js
