+++ assets/js/hooks/password_toggle.js
