+++ assets/js/hooks/virtualized_lines.js
