+++ assets/js/lib/markdown.js
@@ -15,7 +15,7 @@ import { visit } from "unist-util-visit";
 import { toText } from "hast-util-to-text";
 import { removePosition } from "unist-util-remove-position";
 
-import { highlight } from "../cell_editor/live_editor/monaco";
+import { highlight } from "../hooks/cell_editor/live_editor/monaco";
 import { renderMermaid } from "./markdown/mermaid";
 import { escapeHtml } from "../lib/utils";
